BF-69: re-added welcome card
